python-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
openai>=2.30.0
python-dotenv==1.0.1
scikit-learn==1.5.0
pandas==2.2.2
requests==2.32.3
gunicorn==22.0.0
